Reward Handling and DQN Learning
Reward Handling: rewards stored in the Experience tuple are 0 for all intermediate transitions (replaced take_step with drop_ball) and the final +1.0, -1.0, or 0 reward is stored only for the terminal transition.
DQN Learn: our learn function is called based on a trigger (should_learn). This trigger uses a counter of decision steps taken (i.e., times an action was chosen at a ledge or block), this saves us runtime.